@@ -20,7 +20,7 @@
 		}
 	],
 	"require": {
-		"php": ">=5.3.0",
+		"php": ">=5.5",
 		"ext-imap": "*"
 	},
 	"autoload": {
